+++ build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    #[cfg(feature = "napi_support")]
+    napi_build::setup();
+}